+++ even-more-fish-plugin/src/main/java/com/oheers/fish/commands/EMFCommand.java
@@ -10,8 +10,8 @@ import com.oheers.fish.config.MainConfig;
 import com.oheers.fish.config.messages.ConfigMessage;
 import com.oheers.fish.config.messages.Message;
 import com.oheers.fish.config.messages.PrefixType;
-import com.oheers.fish.gui.MainMenuGUI;
-import com.oheers.fish.gui.SellGUI;
+import com.oheers.fish.gui.guis.MainMenuGUI;
+import com.oheers.fish.gui.guis.SellGUI;
 import com.oheers.fish.permissions.AdminPerms;
 import com.oheers.fish.permissions.UserPerms;
 import com.oheers.fish.selling.SellHelper;
