+++ even-more-fish-plugin/src/main/java/com/oheers/fish/gui/GUIUtils.java
@@ -1,14 +1,17 @@
-package com.oheers.fish.utils;
+package com.oheers.fish.gui;
 
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.FishUtils;
 import com.oheers.fish.config.GUIConfig;
 import com.oheers.fish.config.messages.Message;
-import com.oheers.fish.gui.BaitsGUI;
-import com.oheers.fish.gui.EMFGUI;
-import com.oheers.fish.gui.MainMenuGUI;
-import com.oheers.fish.gui.SellGUI;
+import com.oheers.fish.gui.guis.BaitsGUI;
+import com.oheers.fish.gui.guis.EMFGUI;
+import com.oheers.fish.gui.guis.MainMenuGUI;
+import com.oheers.fish.gui.guis.SellGUI;
 import com.oheers.fish.selling.SellHelper;
+import com.oheers.fish.utils.ItemBuilder;
+import com.oheers.fish.utils.ItemFactory;
+import com.oheers.fish.utils.ItemUtils;
 import de.themoep.inventorygui.*;
 import dev.dejvokep.boostedyaml.YamlDocument;
 import dev.dejvokep.boostedyaml.block.implementation.Section;
