+++ even-more-fish-plugin/src/main/java/com/oheers/fish/gui/guis/MainMenuGUI.java
@@ -1,9 +1,9 @@
-package com.oheers.fish.gui;
+package com.oheers.fish.gui.guis;
 
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.config.GUIConfig;
 import com.oheers.fish.config.GUIFillerConfig;
-import com.oheers.fish.utils.GUIUtils;
+import com.oheers.fish.gui.GUIUtils;
 import de.themoep.inventorygui.InventoryGui;
 import dev.dejvokep.boostedyaml.block.implementation.Section;
 import org.bukkit.Material;
