+++ even-more-fish-plugin/src/main/java/com/oheers/fish/gui/guis/SellGUI.java
@@ -1,4 +1,4 @@
-package com.oheers.fish.gui;
+package com.oheers.fish.gui.guis;
 
 import com.github.Anon8281.universalScheduler.scheduling.tasks.MyScheduledTask;
 import com.oheers.fish.EvenMoreFish;
@@ -8,7 +8,7 @@ import com.oheers.fish.config.GUIFillerConfig;
 import com.oheers.fish.config.MainConfig;
 import com.oheers.fish.selling.SellHelper;
 import com.oheers.fish.selling.WorthNBT;
-import com.oheers.fish.utils.GUIUtils;
+import com.oheers.fish.gui.GUIUtils;
 import de.themoep.inventorygui.GuiStorageElement;
 import de.themoep.inventorygui.InventoryGui;
 import dev.dejvokep.boostedyaml.block.implementation.Section;
